% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crabstextfile.R
\name{blue_to_orange}
\alias{blue_to_orange}
\title{Compare two colors together and return the product}
\usage{
blue_to_orange(color_B)
}
\arguments{
\item{x}{Color_O}

\item{y}{Color_B}
}
\value{
The numerical product of x and y
}
\description{
Compare two colors together and return the product
}
