% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mult}
\alias{mult}
\title{Multiply two numbers together and return the product}
\usage{
mult(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
The numerical product of x and y
}
\description{
Multiply two numbers together and return the product
}
\examples{
mult(2, 5)
}
